Extract lints of unit_types group from types group
